refactor: 회원 프로필 API 코드 개선
MemberProfileController의 책임을 분리
